+++ R/expr__string.R
@@ -8,7 +8,7 @@
 #' @description  Parse a Series of dtype Utf8 to a Date/Datetime Series.
 #' @name ExprStr_strptime
 #' @param datatype a temporal data type either pl$Date, pl$Time or pl$Datetime
-#' @param fmt fmt string for parsenig see
+#' @param fmt fmt string for parsing see
 #' see details here https://docs.rs/chrono/latest/chrono/format/strftime/index.html#fn6
 #' Notice time_zone %Z is not supported and will just ignore timezones. Numeric tz  like
 #' %z, %:z  .... are supported.
@@ -17,10 +17,6 @@
 #' @param exact bool , If True, require an exact format match. If False, allow the format to match
 #' anywhere in the target string.
 #' @param cache Use a cache of unique, converted dates to apply the datetime conversion.
-#' @param tz_aware bool, Parse timezone aware datetimes. This may be automatically toggled by the
-#' ‘fmt’ given.
-#' @param utc bool Parse timezone aware datetimes as UTC. This may be useful if you have data with
-#' mixed offsets.
 #' @details Notes When parsing a Datetime the column precision will be inferred from the format
 #' string, if given, eg: “%F %T%.3f” => Datetime(“ms”). If no fractional second component is found
 #' then the default is “us”.
@@ -54,16 +50,14 @@
 #' pl$lit(txt_datetimes)$str$strptime(
 #'   pl$Datetime("ns"),
 #'   fmt = "%Y-%m-%d %H:%M:%S %z", strict = FALSE,
-#'   tz_aware = TRUE, utc = TRUE
 #' )$lit_to_s()
 ExprStr_strptime = function(
     datatype, # : PolarsTemporalType,
     fmt, # : str | None = None,
     strict = TRUE, # : bool = True,
     exact = TRUE, # : bool = True,
-    cache = TRUE, # : bool = True,
-    tz_aware = FALSE, # : bool = False,
-    utc = FALSE) { #-> Expr:
+    cache = TRUE # : bool = True,
+    ) { #-> Expr:
 
   # match on datatype, return Result<Expr, String>
   expr_result = pcase(
@@ -77,7 +71,7 @@ ExprStr_strptime = function(
       tu = .pr$DataType$get_insides(datatype)$tu
 
       .pr$Expr$str_parse_datetime(
-        self, fmt, strict, exact, cache, tz_aware, utc, tu
+        self, fmt, strict, exact, cache, tu
       ) |> and_then(
         \(expr) .pr$Expr$dt_cast_time_unit(expr, tu) # cast if not an err
       )
@@ -198,7 +192,7 @@ ExprStr_to_lowercase = function() {
 #' @keywords ExprStr
 #' @param matches The set of characters to be removed. All combinations of this set of
 #' characters will be stripped. If set to NULL (default), all whitespace is removed instead.
-#' @details will not strip anyt chars beyond the first char not matched. `strip()` starts from
+#' @details will not strip any chars beyond the first char not matched. `strip()` starts from
 #' both left and right. Whereas `lstrip()`and `rstrip()` starts from left and right respectively.
 #' @return Expr of Utf8 lowercase chars
 #' @examples
@@ -219,7 +213,7 @@ ExprStr_strip = function(matches = NULL) {
 #' @keywords ExprStr
 #' @param matches The set of characters to be removed. All combinations of this set of
 #' characters will be stripped. If set to NULL (default), all whitespace is removed instead.
-#' @details will not strip anyt chars beyond the first char not matched. `strip()` starts from
+#' @details will not strip any chars beyond the first char not matched. `strip()` starts from
 #' both left and right. Whereas `lstrip()`and `rstrip()` starts from left and right respectively.
 #' @return Expr of Utf8 lowercase chars
 #' @examples
@@ -240,7 +234,7 @@ ExprStr_lstrip = function(matches = NULL) {
 #' @keywords ExprStr
 #' @param matches The set of characters to be removed. All combinations of this set of
 #' characters will be stripped. If set to NULL (default), all whitespace is removed instead.
-#' @details will not strip anyt chars beyond the first char not matched. `strip()` starts from
+#' @details will not strip any chars beyond the first char not matched. `strip()` starts from
 #' both left and right. Whereas `rstrip()`and `rstrip()` starts from left and right respectively.
 #' @return Expr of Utf8 lowercase chars
 #' @examples
@@ -640,7 +634,7 @@ ExprStr_splitn = function(by, n) {
 #' Replace first matching regex/literal substring with a new string value.
 #' @keywords ExprStr
 #' @param pattern Into<Expr>, regex pattern
-#' @param value Into<Expr> replcacement
+#' @param value Into<Expr> replacement
 #' @param literal bool, Treat pattern as a literal string.
 #'
 #' @return Expr of Utf8 Series
@@ -666,7 +660,7 @@ ExprStr_replace = function(pattern, value, literal = FALSE) {
 #' Replace all matching regex/literal substrings with a new string value.
 #' @keywords ExprStr
 #' @param pattern Into<Expr>, regex pattern
-#' @param value Into<Expr> replcacement
+#' @param value Into<Expr> replacement
 #' @param literal bool, treat pattern as a literal string.
 #'
 #' @return Expr of Utf8 Series
@@ -691,7 +685,7 @@ ExprStr_replace_all = function(pattern, value, literal = FALSE) {
 #' Create subslices of the string values of a Utf8 Series.
 #' @keywords ExprStr
 #' @param pattern Into<Expr>, regex pattern
-#' @param value Into<Expr> replcacement
+#' @param value Into<Expr> replacement
 #' @param literal bool, treat pattern as a literal string.
 #'
 #' @return Expr: Series of dtype Utf8.
@@ -706,19 +700,19 @@ ExprStr_slice = function(offset, length = NULL) {
     unwrap("in str$slice:")
 }
 
-#' explode
-#' @name ExprStr_explode
+#' str_explode
+#' @name ExprStr_str_explode
 #' @aliases expr_str_explode
 #' @description Returns a column with a separate row for every string character.
 #' @keywords ExprStr
 #' @return Expr: Series of dtype Utf8.
 #' @examples
 #' df = pl$DataFrame(a = c("foo", "bar"))
-#' df$select(pl$col("a")$str$explode())
-ExprStr_explode = function() {
-  .pr$Expr$explode(self)
+#' df$select(pl$col("a")$str$str_explode())
+ExprStr_str_explode = function() {
+  .pr$Expr$str_explode(self) |>
+    unwrap("in str$str_explode:")
 }
-
 
 #' parse_int
 #' @name ExprStr_parse_int