+++ R/ipc.R
@@ -8,12 +8,12 @@
 #' @param cache bool, use cache
 #' @param rechunk bool, rechunk reorganize memory layout, potentially make future operations faster, however perform reallocation now.
 #' @param row_count_name NULL or string, if a string add a rowcount column named by this string
-#' @param row_count_offset integer, the rowcount column can be offst by this value
+#' @param row_count_offset integer, the rowcount column can be offset by this value
 #' @param memmap bool, mapped memory
 #'
 #' @return LazyFrame
-#'
-scan_arrow_ipc = function(
+#' @name scan_arrow_ipc
+pl$scan_arrow_ipc = function(
     path, # : str | Path,
     n_rows = NULL, # : int | None = None,
     cache = TRUE, # : bool = True,