+++ R/parquet.R
@@ -4,18 +4,17 @@
 #' @param file string filepath
 #' @param n_rows limit rows to scan
 #' @param cache bool use cache
-#' @param parallel String either Auto, None, Columns or RowGroups. The way to parralize the scan.
+#' @param parallel String either Auto, None, Columns or RowGroups. The way to parallelized the scan.
 #' @param rechunk bool rechunk reorganize memory layout, potentially make future operations faster , however perform reallocation now.
 #' @param row_count_name NULL or string, if a string add a rowcount column named by this string
-#' @param row_count_offset integer, the rowcount column can be offst by this value
+#' @param row_count_offset integer, the rowcount column can be offset by this value
 #' @param low_memory bool, try reduce memory footprint
 #'
 #' @return LazyFrame
-#' @export
-#'
+#' @name scan_parquet
 #' @examples
 #' # TODO write parquet example
-scan_parquet = function(
+pl$scan_parquet = function(
     file, # : str | Path,
     n_rows = NULL, # : int | None = None,
     cache = TRUE, # : bool = True,